requirements.txt
flask
flask_login
# sqlite3
argon2-cffi
cryptography